# docs/swagger.yaml
basePath: /
definitions:
  models.Budget:
    properties:
      budget_identifier:
        enum:
        - flow-engine
        - import-deploy
        type: string
      role:
        type: string
      user_id:
        type: string
      value:
        type: integer
    type: object
info:
  contact: {}
  description: Allows to define budgets and check if a given set of changes would
    violate these budgets.
  title: Budget API
paths:
  /budgets:
    delete:
      description: Deletes a budget by identifier and either user ID or role.
      parameters:
      - description: Budget identifier
        enum:
        - flow-engine
        - import-deploy
        in: query
        name: budget_identifier
        required: true
        type: string
      - description: User ID owning the budget
        in: query
        name: user_id
        type: string
      - description: Role owning the budget
        in: query
        name: role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete budget
      tags:
      - budgets
    get:
      description: Returns a paginated list of configured budgets.
      parameters:
      - default: 100
        description: Maximum number of items to return
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of items to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Budget'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List budgets
      tags:
      - budgets
    put:
      consumes:
      - application/json
      description: Creates a budget or overwrites an existing one for the same identifier
        and subject.
      parameters:
      - description: Budget to create or update
        in: body
        name: budget
        required: true
        schema:
          $ref: '#/definitions/models.Budget'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create or update budget
      tags:
      - budgets
  /check/analytics/flow-engine:
    post:
      consumes:
      - application/json
      description: Validates whether the requested flow-engine operation is within
        the caller budget.
      parameters:
      - description: Original proxied request payload
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "402":
          description: Payment Required
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Check analytics flow-engine budget
      tags:
      - checks
  /check/import/deploy:
    post:
      consumes:
      - application/json
      description: Validates whether the requested import-deploy operation is within
        the caller budget.
      parameters:
      - description: Original proxied request payload
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "402":
          description: Payment Required
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Check import-deploy budget
      tags:
      - checks
  /doc:
    get:
      description: Returns the generated Swagger document for this service.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get OpenAPI document
      tags:
      - documentation
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"